// src/app/generated_services/model/paginationMedicalClearanceDTO.ts
/**
 * BjjClutch Backend API
 * API for managing students, belts, graduation requirements, notices, frequencies and user authentication with Keycloak
 *
 * Contact: marcosturossi@gmail.com.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ShowMedicalClearanceDTO } from './showMedicalClearanceDTO';


export interface PaginationMedicalClearanceDTO { 
    items?: Array<ShowMedicalClearanceDTO> | null;
    totalCount?: number;
    pageNumber?: number;
    pageSize?: number;
    readonly totalPages?: number;
}